--- internal/sherlock/routes_test.go
@@ -1,15 +1,0 @@
-package sherlock
-
-import (
-	"testing"
-
-	"github.com/broadinstitute/sherlock/internal/services"
-)
-
-func TestGetServices(t *testing.T) {
-	// app := New()
-}
-
-type mockDB struct {
-	serviceStore services.MockServiceStore
-}
